refactor ResponseTimeDashboard component to seperate file
